add-struck-project.
Подготовил проект для работы
