+++ StarsWarsInfo/Data/StarsWarsInfo.xcdatamodeld/StarsWarsInfo.xcdatamodel/contents
