+++ StarsWarsInfo/Moduls/MainScreen/MainViewController/MainViewConroller.swift
@@ -1,0 +1,30 @@
+//
+//  MainViewConroller.swift
+//  StarsWarsInfo
+//
+//  Created by Daniil Litvinov on 11.02.2025.
+//
+
+import UIKit
+
+final class MainViewConroller: UIViewController {
+    
+    //MARK: - Life Cicle
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        setupHerarcy()
+        setupLayuot()
+        setupViews()
+    }
+    //MARK: - Setups
+    private func setupHerarcy() {
+    }
+    
+    private func setupViews() {
+        view.backgroundColor = .red
+    }
+    
+    private func setupLayuot() {
+    }
+    
+}
