+++ StarsWarsInfo/Resourse/Assets.xcassets/AccentColor.colorset/Contents.json
