+++ StarsWarsInfo/Resourse/Assets.xcassets/AppIcon.appiconset/Contents.json
