+++ StarsWarsInfo/Resourse/Assets.xcassets/Contents.json
