--- StarsWarsInfo/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  StarsWarsInfo
-//
-//  Created by Daniil Litvinov on 11.02.2025.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
